requirements.txt
pandas==2.2.2
numpy==1.26.4
duckdb==1.0.0
rapidfuzz==3.9.6
python-Levenshtein==0.25.1
phonenumbers==8.13.47
python-dateutil==2.9.0.post0
fastapi==0.112.0
uvicorn==0.30.6
pydantic==2.8.2
streamlit==1.37.1
plotly==5.17.0
transformers==4.36.0
sentence-transformers>=2.6.0,<3.0.0
torch==2.2.2